requirements.txt
Django==4.2.10
mysqlclient==2.2.4
python-dotenv==1.0.1
django-crispy-forms==2.1
crispy-bootstrap5==2024.2
google-auth==2.28.1
google-auth-oauthlib==1.2.0
django-allauth==0.61.1
Pillow==10.2.0
python-dateutil==2.8.2 